Finish W1D coordinator fixup
