+++ wmo/cli/provider_cmd_set_test.py
@@ -1,7 +1,7 @@
 """Provider configuration and pool-registration tests."""
 
 # ruff: noqa: F403, F405
-from wmo.cli.app_fixtures_test import *
+from wmo.cli.cli_fixtures_test import *
 
 
 def test_config_help_does_not_reuse_the_harness_group_name() -> None:
