+++ wmo/repo_docstrings_test.py
@@ -249,6 +249,25 @@ _DOCSTRING_TOMBSTONE_ROWS: Final[tuple[str, ...]] = (
     'wmo/optimize/research/trace_scaling.py|split_mode|function|missing-returns-section',
     'wmo/optimize/routing/scorecard.py|build_ladder|function|missing-args-section',
     'wmo/optimize/routing/scorecard.py|build_ladder|function|missing-returns-section',
+    'wmo/optimize/routing/scorecard.py|build_scorecard|function|missing-returns-section',
+    'wmo/optimize/routing/scorecard.py|CompletionRule.completed|method|missing-args-section',
+    'wmo/optimize/routing/scorecard.py|CompletionRule.completed|method|missing-returns-section',
+    ('wmo/optimize/routing/scorecard.py|CompletionRule.completed|method|'
+     'nontrivial-one-line-docstring'),
+    'wmo/optimize/routing/scorecard.py|ConditionLabel.replace|method|missing-args-section',
+    'wmo/optimize/routing/scorecard.py|ConditionLabel.replace|method|missing-returns-section',
+    'wmo/optimize/routing/scorecard.py|Ladder.operating_points|method|missing-args-section',
+    'wmo/optimize/routing/scorecard.py|Ladder.operating_points|method|missing-returns-section',
+    ('wmo/optimize/routing/scorecard.py|Ladder.operating_points|method|'
+     'nontrivial-one-line-docstring'),
+    'wmo/optimize/routing/scorecard.py|Ladder.pareto|method|missing-args-section',
+    'wmo/optimize/routing/scorecard.py|Ladder.pareto|method|missing-returns-section',
+    'wmo/optimize/routing/scorecard.py|OperatingPoint.as_cost_quality_anchor|method|missing-returns-section',
+    'wmo/optimize/routing/scorecard.py|rows_for_model|function|missing-args-section',
+    'wmo/optimize/routing/scorecard.py|rows_for_model|function|missing-raises-section',
+    'wmo/optimize/routing/scorecard.py|rows_for_model|function|missing-returns-section',
+    ('wmo/optimize/routing/scorecard.py|rows_for_model|function|'
+     'nontrivial-one-line-docstring'),
     'wmo/optimize/routing/scorecard.py|rows_for_policy|function|missing-returns-section',
     'wmo/optimize/telemetry/hooks.py|GridEmitter.create|method|missing-returns-section',
     'wmo/optimize/telemetry/hooks.py|GridEmitter.on_arm_start|method|missing-args-section',
